Fixing user id from long to int

+++ gamingMatchMaker/src/main/java/com/gamingMatchMaker/gamingMatchMaker/dao/UserRepository.java
@@ -8,5 +8,5 @@ import java.util.UUID;
 
 public interface UserRepository extends JpaRepository<UserRec, UUID> {
     Optional<UserRec> findByEmail(String email);
-    Optional<UserRec> findById(long id);
+    Optional<UserRec> findById(int id);
 }
